:recycle: [repositories] Move tests for `registry` to separate module
